QtSensors: Fix compilation of examples/warnings on Windows.
- Fix double->float truncation in tests
- Use -rdynamic for g++ only
- Introduce EXPORT macro in grue library for Windows to work
- Add DESTDIR for the grue lib (to prevent it from ending up
  in the debug/release subfolders, respectively

Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>

+++ tests/auto/sensors2qmlapi/qdeclaccelerometer.h
@@ -65,13 +65,13 @@ public:
     }
     bool isActive() { return _active; }
 
-    void test(float x, float y, float z)
+    void test(qreal x, qreal y, qreal z)
     {
         if (sensor()->filters().count() > 0){
             QAccelerometerFilter* af = (QAccelerometerFilter*)sensor()->filters().at(0);
-            reader.setX((qreal)x);
-            reader.setY((qreal)y);
-            reader.setZ((qreal)z);
+            reader.setX(x);
+            reader.setY(y);
+            reader.setZ(z);
             af->filter(&reader);
         }
     }
